Completed work on basic Homework submission API and client side functionality.
Homework will be submitted as array of answer objects including question and answer strings.
The result of the API call will be an array of the same objects including a correct boolean that will be true if the question was answered correctly.

When submitting homework, the amounts homework has been submitted and the amount of times a certain answer has been given will be logged in the database.