Finished work on homework submissions. Submission results will be saved for every submission made in the database. Homework can only be submitted once.
When requesting all homework with the student API, the returned object will contain available and completed Arrays of homework.

When requesting a single piece of homework with the student API, the returned object will contain a submission object including the submission details if the homework has already been done by the student.